cli(refactor): refine update changefeed cmd
